refactor: align product search request builder naming
