+++ packages/web-components/src/search/index.ts
@@ -2,7 +2,7 @@ export * from './components';
 export * from './enums'; 
 export * from './product-search-overlay';
 export * from './product-search';
-export * from './productSearchRequest';
+export * from './productSearchRequestBuilder';
 export * from './searchSortingBuilder';
 export * from './searcher'; 
 export * from './types';
